extract remmaper function
